+++ CHANGELOG.md
@@ -10,6 +10,9 @@ Current
 
 ### Added:
 
+- [Add custom dimension row serializer](https://github.com/yahoo/fili/pull/399)
+    * Add custom dimension row serializer so that dimension row serialize into dimension row cache correctly
+
 - [Fili-security module](https://github.com/yahoo/fili/pull/405)
     * Added security module for fili data security filters
     * Created `ChainingRequestMapper`, and a set of mappers for gatekeeping on security roles and whitelisting dimension filters.